moved download file from simulations
