+++ datasetinsights/data/exceptions.py
@@ -1,0 +1,3 @@
+class DownloadError(Exception):
+    """ Raise when download file failed.
+    """
