+++ datasetinsights/data/simulation/exceptions.py
@@ -1,8 +1,3 @@
-class DownloadError(Exception):
-    """ Raise when download file failed.
-    """
-
-
 class DefinitionIDError(Exception):
     """ Raise when a given definition id can't be found.
     """
